Conflicts:
	Sample/source/BaseFrame.cpp
	Sample/source/Demo.cpp
	Sample/source/ShaderGroup.cpp

Former-commit-id: 264a5510daca37861e9cff119e8e33c8c132e052

+++ Sample/includes/json/json/json.h
@@ -68,6 +68,8 @@ license you like.
 // End of content of file: LICENSE
 // //////////////////////////////////////////////////////////////////////
 
+#include <cstdint>
+
 //This define was missing for some reason, I added it
 #define JSON_IS_AMALGAMATION
 
@@ -167,8 +169,8 @@ namespace Json {
    typedef __int64 Int64;
    typedef unsigned __int64 UInt64;
 #  else // if defined(_MSC_VER) // Other platforms, use long long
-   typedef long long int Int64;
-   typedef unsigned long long int UInt64;
+   typedef int64_t Int64;
+   typedef uint64_t UInt64;
 #  endif // if defined(_MSC_VER)
    typedef Int64 LargestInt;
    typedef UInt64 LargestUInt;
